requirements.txt
Django == 5.0.6
djangorestframework
djangorestframework-api-key
gunicorn